refactor(server): Removed the old CertificateVerification implementation
