+++ examples/pubsub/sks/server_pubsub_central_sks.c
@@ -251,9 +251,7 @@ main(int argc, char **argv) {
     UA_ByteString revocationList[100];
     size_t revocationListSize = 0;
 #else
-    const char *trustlistFolder = NULL;
-    const char *issuerlistFolder = NULL;
-    const char *revocationlistFolder = NULL;
+    char *pkiFolder = NULL;
 #endif /* __linux__ */
 
 #endif /* UA_ENABLE_ENCRYPTION */
@@ -362,33 +360,13 @@ main(int argc, char **argv) {
             continue;
         }
 #else  /* __linux__ */
-        if(strcmp(argv[pos], "--trustlistFolder") == 0) {
+        if(strcmp(argv[pos], "--pkiFolder") == 0) {
             filetype = 't';
             continue;
         }
 
-        if(strcmp(argv[pos], "--issuerlistFolder") == 0) {
-            filetype = 'l';
-            continue;
-        }
-
-        if(strcmp(argv[pos], "--revocationlistFolder") == 0) {
-            filetype = 'r';
-            continue;
-        }
-
         if(filetype == 't') {
-            trustlistFolder = argv[pos];
-            continue;
-        }
-
-        if(filetype == 'l') {
-            issuerlistFolder = argv[pos];
-            continue;
-        }
-
-        if(filetype == 'r') {
-            revocationlistFolder = argv[pos];
+            pkiFolder = argv[pos];
             continue;
         }
 #endif /* __linux__ */
@@ -408,26 +386,16 @@ main(int argc, char **argv) {
         issuerListSize, revocationList, revocationListSize);
     if(res != UA_STATUSCODE_GOOD)
         goto cleanup;
-#else /* On Linux we can monitor the certs folder and reload when changes are made */
-    UA_StatusCode res = UA_ServerConfig_setDefaultWithSecurityPolicies(
-        &config, port, &certificate, &privateKey, NULL, 0, NULL, 0, NULL, 0);
-    if(res != UA_STATUSCODE_GOOD)
-        goto cleanup;
-#ifdef UA_ENABLE_CERT_REJECTED_DIR
-    res |= UA_CertificateVerification_CertFolders(&config.secureChannelPKI,
-                                                  trustlistFolder, issuerlistFolder,
-                                                  revocationlistFolder, NULL);
-    res |= UA_CertificateVerification_CertFolders(&config.sessionPKI,
-                                                 trustlistFolder, issuerlistFolder,
-                                                 revocationlistFolder, NULL);
-#else
-    res |= UA_CertificateVerification_CertFolders(&config.secureChannelPKI,
-                                                  trustlistFolder, issuerlistFolder,
-                                                  revocationlistFolder);
-    res |= UA_CertificateVerification_CertFolders(&config.sessionPKI,
-                                                  trustlistFolder, issuerlistFolder,
-                                                  revocationlistFolder);
-#endif
+#else /* On Linux we can monitor the pki folder and reload when changes are made */
+    UA_String *pkiStoreFolder = NULL;
+    if(pkiFolder) {
+        pkiStoreFolder = UA_String_new();
+        *pkiStoreFolder = UA_STRING_ALLOC(pkiFolder);
+    }
+    UA_StatusCode res = UA_ServerConfig_setDefaultWithFilestore(
+        &config, port, &certificate, &privateKey, pkiStoreFolder);
+    if(pkiStoreFolder)
+        UA_String_clear(pkiStoreFolder);
     if(res != UA_STATUSCODE_GOOD)
         goto cleanup;
 #endif /* __linux__ */
