+++ plugins/crypto/ua_certificategroup_none.c
@@ -20,7 +20,7 @@ clearVerifyAllowAll(UA_CertificateGroup *certGroup) {
 
 }
 
-void UA_CertificateVerification_AcceptAll(UA_CertificateGroup *certGroup) {
+void UA_CertificateGroup_AcceptAll(UA_CertificateGroup *certGroup) {
     /* Clear the structure, as it may have already been initialized. */
     UA_NodeId groupId = certGroup->certificateGroupId;
     if(certGroup->clear)
