+++ plugins/ua_config_json.c
@@ -759,80 +759,54 @@ PARSE_JSON(SecurityPolciesField) {
 
 PARSE_JSON(SecurityPkiField) {
 #ifdef UA_ENABLE_ENCRYPTION
-    UA_CertificateGroup *field = (UA_CertificateGroup*)configField;
-    UA_String trustListFolder = {.length = 0, .data = NULL};
-    UA_String issuerListFolder = {.length = 0, .data = NULL};
-    UA_String revocationListFolder = {.length = 0, .data = NULL};
+    UA_ServerConfig *config = (UA_ServerConfig*)configField;
+    UA_String pkiFolder = {.length = 0, .data = NULL};
 
     cj5_token tok = ctx->tokens[++ctx->index];
-    for(size_t i = tok.size/2; i > 0; i--) {
-        tok = ctx->tokens[++ctx->index];
-        switch(tok.type) {
-        case CJ5_TOKEN_STRING: {
-            char *field_str = (char*)UA_malloc(tok.size + 1);
-            unsigned int str_len = 0;
-            cj5_get_str(&ctx->result, (unsigned int)ctx->index, field_str, &str_len);
-            if(strcmp(field_str, "trustListFolder") == 0) {
-                parseJsonJumpTable[UA_SERVERCONFIGFIELD_STRING](ctx, &trustListFolder, NULL);
-            } else if(strcmp(field_str, "issuerListFolder") == 0) {
-                parseJsonJumpTable[UA_SERVERCONFIGFIELD_STRING](ctx, &issuerListFolder, NULL);
-            } else if(strcmp(field_str, "revocationListFolder") == 0) {
-                parseJsonJumpTable[UA_SERVERCONFIGFIELD_STRING](ctx, &revocationListFolder, NULL);
-            } else {
-                UA_LOG_ERROR(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND, "Unknown field name.");
-            }
-            UA_free(field_str);
-            break;
-        }
-        default:
-            break;
-        }
-    }
+    UA_ByteString buf = getJsonPart(tok, ctx->json);
+    UA_StatusCode retval = UA_decodeJson(&buf, &pkiFolder, &UA_TYPES[UA_TYPES_STRING], NULL);
+    if(retval != UA_STATUSCODE_GOOD)
+        return retval;
+
 #ifndef __linux__
     /* Currently not supported! */
-    (void)field;
+    (void)config;
     return UA_STATUSCODE_GOOD;
 #else
-    /* set server config field */
-    char *sTrustListFolder = NULL;
-    char *sIssuerListFolder = NULL;
-    char *sRevocationListFolder = NULL;
-    if(trustListFolder.length > 0) {
-        sTrustListFolder = (char*)UA_malloc(trustListFolder.length+1);
-        memcpy(sTrustListFolder, trustListFolder.data, trustListFolder.length);
-        sTrustListFolder[trustListFolder.length] = '\0';
-    }
-    if(issuerListFolder.length > 0) {
-        sIssuerListFolder = (char*)UA_malloc(issuerListFolder.length+1);
-        memcpy(sIssuerListFolder, issuerListFolder.data, issuerListFolder.length);
-        sIssuerListFolder[issuerListFolder.length] = '\0';
-    }
-    if(revocationListFolder.length > 0) {
-        sRevocationListFolder = (char*)UA_malloc(revocationListFolder.length+1);
-        memcpy(sRevocationListFolder, revocationListFolder.data, revocationListFolder.length);
-        sRevocationListFolder[revocationListFolder.length] = '\0';
-    }
-    if(field && field->clear)
-        field->clear(field);
-#ifdef UA_ENABLE_CERT_REJECTED_DIR
-    UA_StatusCode retval = UA_CertificateVerification_CertFolders(field, sTrustListFolder,
-                                                                  sIssuerListFolder, sRevocationListFolder, NULL);
-#else
-    UA_StatusCode retval = UA_CertificateVerification_CertFolders(field, sTrustListFolder,
-                                                                  sIssuerListFolder, sRevocationListFolder);
-#endif
-    /* Clean up */
-    if(sTrustListFolder)
-        UA_free(sTrustListFolder);
-    if(sIssuerListFolder)
-        UA_free(sIssuerListFolder);
-    if(sRevocationListFolder)
-        UA_free(sRevocationListFolder);
-    UA_String_clear(&trustListFolder);
-    UA_String_clear(&issuerListFolder);
-    UA_String_clear(&revocationListFolder);
+    /* Set up the parameters */
+    UA_KeyValuePair params[2];
+    size_t paramsSize = 2;
 
-    return retval;
+    params[0].key = UA_QUALIFIEDNAME(0, "max-trust-listsize");
+    UA_Variant_setScalar(&params[0].value, &config->maxTrustListSize, &UA_TYPES[UA_TYPES_UINT32]);
+    params[1].key = UA_QUALIFIEDNAME(0, "max-rejected-listsize");
+    UA_Variant_setScalar(&params[1].value, &config->maxRejectedListSize, &UA_TYPES[UA_TYPES_UINT32]);
+
+    UA_KeyValueMap paramsMap;
+    paramsMap.map = params;
+    paramsMap.mapSize = paramsSize;
+
+    /* set server config field */
+    UA_NodeId defaultApplicationGroup =
+           UA_NODEID_NUMERIC(0, UA_NS0ID_SERVERCONFIGURATION_CERTIFICATEGROUPS_DEFAULTAPPLICATIONGROUP);
+    retval = UA_CertificateGroup_Filestore(&config->secureChannelPKI, &defaultApplicationGroup,
+                                           &pkiFolder, config->logging, &paramsMap);
+    if(retval != UA_STATUSCODE_GOOD) {
+        UA_String_clear(&pkiFolder);
+        return retval;
+    }
+
+    UA_NodeId defaultUserTokenGroup =
+            UA_NODEID_NUMERIC(0, UA_NS0ID_SERVERCONFIGURATION_CERTIFICATEGROUPS_DEFAULTUSERTOKENGROUP);
+    retval = UA_CertificateGroup_Filestore(&config->sessionPKI, &defaultUserTokenGroup,
+                                            &pkiFolder, config->logging, &paramsMap);
+    if(retval != UA_STATUSCODE_GOOD) {
+        UA_String_clear(&pkiFolder);
+        return retval;
+    }
+
+    /* Clean up */
+    UA_String_clear(&pkiFolder);
 #endif
 #endif
     return UA_STATUSCODE_GOOD;
@@ -1010,10 +984,8 @@ parseJSONConfig(UA_ServerConfig *config, UA_ByteString json_config) {
 #ifdef UA_ENABLE_ENCRYPTION
                 else if(strcmp(field, "securityPolicies") == 0)
                     retval = parseJsonJumpTable[UA_SERVERCONFIGFIELD_SECURITYPOLICIES](&ctx, config, NULL);
-                else if(strcmp(field, "secureChannelPKI") == 0)
-                    retval = parseJsonJumpTable[UA_SERVERCONFIGFIELD_SECURITYPKI](&ctx, &config->secureChannelPKI, NULL);
-                else if(strcmp(field, "sessionPKI") == 0)
-                    retval = parseJsonJumpTable[UA_SERVERCONFIGFIELD_SECURITYPKI](&ctx, &config->sessionPKI, NULL);
+                else if(strcmp(field, "pkiFolder") == 0)
+                    retval = parseJsonJumpTable[UA_SERVERCONFIGFIELD_SECURITYPKI](&ctx, config, NULL);
 #endif
                 else {
                     UA_LOG_WARNING(UA_Log_Stdout, UA_LOGCATEGORY_USERLAND, "Field name '%s' unknown or misspelled. Maybe the feature is not enabled either.", field);
