+++ tests/client/check_client_authentication.c
@@ -69,7 +69,7 @@ static void setup(void) {
     ck_assert(server != NULL);
 
     UA_ServerConfig *config = UA_Server_getConfig(server);
-    UA_CertificateVerification_AcceptAll(&config->secureChannelPKI);
+    UA_CertificateGroup_AcceptAll(&config->secureChannelPKI);
 
     UA_Server_run_startup(server);
     THREAD_CREATE(server_thread, serverloop);
@@ -110,7 +110,7 @@ START_TEST(Client_connect_certificate) {
 
     UA_ClientConfig_setDefaultEncryption(cc, certificate, privateKey,
                                          NULL, 0, NULL, 0);
-    UA_CertificateVerification_AcceptAll(&cc->certificateVerification);
+    UA_CertificateGroup_AcceptAll(&cc->certificateVerification);
 
     /* Set the ApplicationUri used in the certificate */
     UA_String_clear(&cc->clientDescription.applicationUri);
@@ -154,7 +154,7 @@ START_TEST(Client_connect_invalid_certificate) {
 
         UA_ClientConfig_setDefaultEncryption(cc, certificate, privateKey,
                                              NULL, 0, NULL, 0);
-        UA_CertificateVerification_AcceptAll(&cc->certificateVerification);
+        UA_CertificateGroup_AcceptAll(&cc->certificateVerification);
 
         /* Set the ApplicationUri used in the certificate */
         UA_String_clear(&cc->clientDescription.applicationUri);
