+++ tests/server/check_discovery.c
@@ -69,8 +69,8 @@ configure_lds_server(UA_Server *pServer) {
                                                    NULL, 0, NULL, 0, NULL, 0);
     config_lds->tcpReuseAddr = true;
 
-    UA_CertificateVerification_AcceptAll(&config_lds->secureChannelPKI);
-    UA_CertificateVerification_AcceptAll(&config_lds->sessionPKI);
+    UA_CertificateGroup_AcceptAll(&config_lds->secureChannelPKI);
+    UA_CertificateGroup_AcceptAll(&config_lds->sessionPKI);
 
     config_lds->applicationDescription.applicationType =
         UA_APPLICATIONTYPE_DISCOVERYSERVER;
@@ -154,8 +154,8 @@ setup_register(void) {
 
     config_register->tcpReuseAddr = true;
 
-    UA_CertificateVerification_AcceptAll(&config_register->secureChannelPKI);
-    UA_CertificateVerification_AcceptAll(&config_register->sessionPKI);
+    UA_CertificateGroup_AcceptAll(&config_register->secureChannelPKI);
+    UA_CertificateGroup_AcceptAll(&config_register->sessionPKI);
 
     UA_String_clear(&config_register->applicationDescription.applicationUri);
     config_register->applicationDescription.applicationUri =
@@ -194,7 +194,7 @@ registerServer(void) {
     UA_ClientConfig cc;
     memset(&cc, 0, sizeof(UA_ClientConfig));
     UA_ClientConfig_setDefaultEncryption(&cc, certificate, privateKey, NULL, 0, NULL, 0);
-    UA_CertificateVerification_AcceptAll(&cc.certificateVerification);
+    UA_CertificateGroup_AcceptAll(&cc.certificateVerification);
     cc.eventLoop->dateTime_now = UA_DateTime_now_fake;
     cc.eventLoop->dateTime_nowMonotonic = UA_DateTime_now_fake;
 
@@ -225,7 +225,7 @@ unregisterServer(void) {
     UA_ClientConfig cc;
     memset(&cc, 0, sizeof(UA_ClientConfig));
     UA_ClientConfig_setDefaultEncryption(&cc, certificate, privateKey, NULL, 0, NULL, 0);
-    UA_CertificateVerification_AcceptAll(&cc.certificateVerification);
+    UA_CertificateGroup_AcceptAll(&cc.certificateVerification);
     cc.eventLoop->dateTime_now = UA_DateTime_now_fake;
     cc.eventLoop->dateTime_nowMonotonic = UA_DateTime_now_fake;
 
@@ -270,7 +270,7 @@ Server_register_semaphore(void) {
     UA_ClientConfig cc;
     memset(&cc, 0, sizeof(UA_ClientConfig));
     UA_ClientConfig_setDefaultEncryption(&cc, certificate, privateKey, NULL, 0, NULL, 0);
-    UA_CertificateVerification_AcceptAll(&cc.certificateVerification);
+    UA_CertificateGroup_AcceptAll(&cc.certificateVerification);
     cc.eventLoop->dateTime_now = UA_DateTime_now_fake;
     cc.eventLoop->dateTime_nowMonotonic = UA_DateTime_now_fake;
